Move all Shadcn files to consolidated folder
